# test/api/env/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IIA Server API",
    "description": "IIA (Intelligent Information Assistant) Server API Documentation",
    "contact": {
      "name": "Charles",
      "email": "inforassistant@foxmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9424",
      "description": "Local Development Environment"
    },
    {
      "url": "http://10.15.0.21:9424",
      "description": "Test Environment"
    }
  ],
  "security": [
    {
      "bearer-jwt": []
    }
  ],
...
  "paths": {
    "/api/reminder/tag/update": {
      "put": {
        "tags": [
          "Tag Management"
        ],
        "summary": "Update Tag",
        "description": "Update tag information including name and color",
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tag updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "404": {
            "description": "Tag not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
...
    "/api/reminder/tag/delete": {
      "post": {
        "tags": [
          "Tag Management"
        ],
        "summary": "Delete Tag",
        "description": "Delete a tag by ID",
        "operationId": "delete_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagDeleteDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tag deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "404": {
            "description": "Tag not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
    "/api/reminder/tag/create": {
      "post": {
        "tags": [
          "Tag Management"
        ],
        "summary": "Create Tag",
        "description": "Create a new tag with name and color",
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tag created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
...
    "/api/reminder/tag/get-all": {
      "get": {
        "tags": [
          "Tag Management"
        ],
        "summary": "Get All Tags",
        "description": "Get all tags for the current user",
        "operationId": "getAll_1",
        "responses": {
          "200": {
            "description": "Tags retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsListTag"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsListTag"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "TagUpdateDTO": {
        "type": "object",
        "description": "Tag update request data",
        "properties": {
          "tagId": {
            "type": "integer",
            "format": "int64",
            "description": "Tag ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Tag name",
            "example": "Work",
            "maxLength": 20,
            "minLength": 1,
            "pattern": "^(?!\\s*$).+"
          },
          "color": {
            "type": "string",
            "description": "Tag color",
            "example": "#409EFF"
          }
        },
        "required": [
          "color",
          "name",
          "tagId"
        ]
      },
      "ResponseUtilsVoid": {
        "type": "object",
        "description": "Standard API response structure",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "Response status code",
            "example": 200,
            "maximum": 599,
            "minimum": 100
          },
          "msg": {
            "type": "string",
            "description": "Response message",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "Response data payload, can be null for error responses"
          }
        }
      },
...
      "TagDeleteDTO": {
        "type": "object",
        "description": "Tag delete request data",
        "properties": {
          "tagId": {
            "type": "integer",
            "format": "int64",
            "description": "Tag ID",
            "example": 1
          }
        },
        "required": [
          "tagId"
        ]
      },
      "TagCreateDTO": {
        "type": "object",
        "description": "Tag create request data",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tag name",
            "example": "Work",
            "maxLength": 20,
            "minLength": 1
          },
          "color": {
            "type": "string",
            "description": "Tag color",
            "example": "#409EFF"
          }
        },
        "required": [
          "color",
          "name"
        ]
      },
...
      "ResponseUtilsListTag": {
        "type": "object",
        "description": "Standard API response structure",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "Response status code",
            "example": 200,
            "maximum": 599,
            "minimum": 100
          },
          "msg": {
            "type": "string",
            "description": "Response message",
            "example": "success"
          },
          "data": {
            "type": "array",
            "description": "Response data payload, can be null for error responses",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          }
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "tagId": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "operationId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "type": "http",
        "description": "JWT authentication token, format: Bearer {token}",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "x-apifox": {
          "variable": "access_token"
        }
      }
    }
  }
}